feat(server): persist and restore player inventory
Store player inventory in player storage data so it survives sessions, and send inventory to the client on login. Add position tracking to inventory stacks for slot placement, rename ItemStack sum to count, and add bounds checks when modifying inventory slots.

+++ include/Cubed/gameplay/client/local_player.hpp
@@ -81,7 +81,7 @@ public:
 
     int selected_hotbar() const;
 
-    void set_inventory(int pos, std::optional<ItemStack> item);
+    void set_inventory(size_t pos, std::optional<ItemStack> item);
 
     std::span<const std::optional<ItemStack>> get_inventory() const;
